Extract code related to data-flow analysis into `dfa` package
