+++ src/main/kotlin/org/rust/ide/highlight/RsHighlightExitPointsHandlerFactory.kt
@@ -12,7 +12,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.util.Consumer
-import org.rust.lang.core.cfg.ExitPoint
+import org.rust.lang.core.dfa.ExitPoint
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.RsElementTypes.*
 import org.rust.lang.core.psi.ext.*
