+++ src/main/kotlin/org/rust/ide/inspections/RsExtraSemicolonInspection.kt
@@ -9,7 +9,7 @@ import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElementVisitor
-import org.rust.lang.core.cfg.ExitPoint
+import org.rust.lang.core.dfa.ExitPoint
 import org.rust.lang.core.psi.RsExprStmt
 import org.rust.lang.core.psi.RsFunction
 import org.rust.lang.core.psi.RsVisitor
