+++ src/main/kotlin/org/rust/lang/core/dfa/borrowck/LoanPath.kt
@@ -3,16 +3,16 @@
  * found in the LICENSE file.
  */
 
-package org.rust.lang.core.types.borrowck
+package org.rust.lang.core.dfa.borrowck
 
+import org.rust.lang.core.dfa.Categorization
+import org.rust.lang.core.dfa.Cmt
+import org.rust.lang.core.dfa.MutabilityCategory
+import org.rust.lang.core.dfa.PointerKind
+import org.rust.lang.core.dfa.borrowck.LoanPathElement.Deref
+import org.rust.lang.core.dfa.borrowck.LoanPathKind.*
 import org.rust.lang.core.psi.RsPatBinding
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.types.borrowck.LoanPathElement.Deref
-import org.rust.lang.core.types.borrowck.LoanPathKind.*
-import org.rust.lang.core.types.infer.Categorization
-import org.rust.lang.core.types.infer.Cmt
-import org.rust.lang.core.types.infer.MutabilityCategory
-import org.rust.lang.core.types.infer.PointerKind
 import org.rust.lang.core.types.regions.Scope
 import org.rust.lang.core.types.ty.Ty
 import java.util.*
