+++ src/main/kotlin/org/rust/lang/core/types/Extensions.kt
@@ -12,15 +12,21 @@ import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider.Result
 import com.intellij.psi.util.CachedValuesManager
 import com.intellij.psi.util.PsiModificationTracker
-import org.rust.lang.core.cfg.ControlFlowGraph
+import org.rust.lang.core.dfa.Cmt
+import org.rust.lang.core.dfa.ControlFlowGraph
+import org.rust.lang.core.dfa.MemoryCategorizationContext
+import org.rust.lang.core.dfa.borrowck.BorrowCheckContext
+import org.rust.lang.core.dfa.borrowck.BorrowCheckResult
+import org.rust.lang.core.dfa.liveness.Liveness
+import org.rust.lang.core.dfa.liveness.LivenessContext
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.ext.*
 import org.rust.lang.core.resolve.ImplLookup
 import org.rust.lang.core.resolve.KnownItems
 import org.rust.lang.core.resolve.knownItems
-import org.rust.lang.core.types.borrowck.BorrowCheckContext
-import org.rust.lang.core.types.borrowck.BorrowCheckResult
-import org.rust.lang.core.types.infer.*
+import org.rust.lang.core.types.infer.RsInferenceResult
+import org.rust.lang.core.types.infer.inferTypeReferenceType
+import org.rust.lang.core.types.infer.inferTypesIn
 import org.rust.lang.core.types.regions.getRegionScopeTree
 import org.rust.lang.core.types.ty.Ty
 import org.rust.lang.core.types.ty.TyTypeParameter
